feat: update and delete by id

+++ src/players/services/players.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreatePlayerDto } from '../dtos/create-player.dto';
 import { Player } from '../models/player.schema';
 import { PlayersRepository } from '../repositories/players.repository';
@@ -9,6 +9,12 @@ export class PlayersService {
   constructor(private readonly playersRepository: PlayersRepository) {}
 
   async create(createPlayerDto: CreatePlayerDto): Promise<Player> {
+    const findByEmail = this.playersRepository.findByEmail(createPlayerDto.email);
+
+    if (findByEmail) {
+      throw new BadRequestException('e-mail already in use');
+    }
+
     return this.playersRepository.create(createPlayerDto);
   }
 
@@ -17,10 +23,22 @@ export class PlayersService {
   }
 
   async findById(id: string): Promise<Player> {
-    return this.playersRepository.findById(id);
+    const player = this.playersRepository.findById(id);
+
+    if (!player) {
+      throw new NotFoundException();
+    }
+
+    return player;
   }
 
   async deleteById(id: string) {
+    const player = this.playersRepository.findById(id);
+
+    if (!player) {
+      throw new NotFoundException();
+    }
+
     return this.playersRepository.delete(id);
   }
